+++ ws-server/src/future.rs
@@ -160,7 +160,7 @@ struct MonitorInner {
 
 /// Monitor for checking whether the server has been flagged to shut down.
 #[derive(Debug, Clone)]
-pub struct StopMonitor(Arc<MonitorInner>);
+pub(crate) struct StopMonitor(Arc<MonitorInner>);
 
 impl Drop for StopMonitor {
 	fn drop(&mut self) {
