+++ ws-server/src/lib.rs
@@ -34,7 +34,6 @@ extern crate alloc;
 
 mod future;
 mod server;
-pub mod tmp;
 
 #[cfg(test)]
 mod tests;
